Added CRUD operation for Inventory movement

+++ src/CKMS/CKMS.Interfaces/Repository/IGenericRepository.cs
@@ -3,7 +3,7 @@
     public interface IGenericRepository<TEntity> where TEntity : class
     {
         Task AddAsync(TEntity entity);
-        Task<TEntity?> GetByIdAsync(int id);
+        Task<TEntity?> GetByIdAsync(Int64 id);
         Task<TEntity?> GetByGuidAsync(Guid id);
         Task<List<TEntity?>> GetAllAsync(bool tracked = false);
         void Update(TEntity entity);
